refractor functions of to_dot_graph
